Fix llama_stack_client log levels
Log levels on the llama_stack_client are currently broken and setting
`LLAMA_STACK_LOG=debug` doesn't work as it should.  Log levels are
automatically changed to WARNING.

This is caused by us blindly loading the `llama_stack` library, which
initializes logging to WARNING except a subset of modules that can be
modified using the `LLAMA_STACK_LOGGING` env var.

With this patch we only import `llama_stack` when we are actually going
to use it, so that when we use the normal llama-stack client we can set
its logging level.