Allow running a subset of unit or integration tests
Add support for filtering unit and integration tests using an
expression.

This allow for running only tests whose names or parent class names
match a given expression. For example, to run tests that matches the word
"endpoint", do:

$ make test-unit TEST_FILTER=endpoint

Or, to run tests that matches the work "endpoint" and "utils" do:

$ make test-unit TEST_FILTER="endpoint or utils"

Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>