Delay setup_model_metrics() until /metrics is called
This patch moves the setup_model_metrics() from the service startup to
the first time the /metrics endpoint is called. This speed up the
lightspeed-stack service initialization and also make lightspeed-stack
more resilient regarding service initialization order because it no
longer requires llama-stack to be started first (as
setup_model_metrics() tries to connect to llama-stack and fetch the list
of models from it).

Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>